requirements.txt
# Core dependencies
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# Deep Learning
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.30.0
lightning>=2.0.0

# EEG Processing
mne>=1.0.0

# Data Storage
h5py>=3.0.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Utilities
tqdm>=4.62.0
pyyaml>=6.0
omegaconf>=2.3.0
python-Levenshtein>=0.21.0
sentencepiece>=0.1.99

# Development & Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-xdist>=3.0.0
pytest-timeout>=2.1.0
pytest-benchmark>=4.0.0
black>=22.0.0
flake8>=4.0.0
isort>=5.10.0
mypy>=1.0.0
radon>=5.1.0
pylint>=2.17.0
safety>=2.3.0
bandit>=1.7.0
coverage>=7.0.0

# Experiment Tracking
tensorboard>=2.13.0
wandb>=0.15.0

# Documentation
sphinx>=5.0.0
sphinx-rtd-theme>=1.2.0
